fix: use offset-based mysql timezone when writing field values
When MySQL's named timezone tables are not available, an offset-based timezone is applied to the session. The downside of offset-based timezones though is they do not adjust for daylight savings differences between dates. When the database connection is using an offset, timezones sent to the database must be rendered with this same offset